Refactor Loader imports for consistency and enhance loading state management in FormContainer
