+++ app/[locale]/loading.tsx
@@ -1,4 +1,4 @@
-import { Loader } from '@/components/loader/loader'
+import { Loader } from '@/components'
 
 const Loading = () => {
   return <Loader />
