+++ components/index.ts
@@ -22,3 +22,4 @@ export * from './partners-list/PartnersList'
 export * from './photo-gallery/PhotoGallery'
 export * from './rich-text/RichText'
 export * from './sponsors/Sponsors'
+export * from './loader/loader'
